# .mem/openapi.yaml
openapi: 3.0.3
info:
  title: Co-Work API
  description: API documentation for the Co-Work collaborative project management system
  version: 1.0.0
  contact:
    name: Co-Work Team
servers:
  - url: /api
    description: API base path
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required:
        - error
...
    Message:
      type: object
      properties:
        id:
          type: integer
          description: Message ID
        projectId:
          type: integer
          description: Project ID
        role:
          type: string
          enum: [user, assistant]
          description: Message sender role
        content:
          type: string
          description: Message content
        quotedMessageId:
          type: integer
          nullable: true
          description: Quoted message ID
        pinned:
          type: boolean
          description: Pinned status
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - projectId
        - role
        - content
        - pinned
        - createdAt
    MessageContextFile:
      type: object
      properties:
        id:
          type: integer
          description: Context file association ID
        messageId:
          type: integer
          description: Message ID
        fileId:
          type: integer
          description: File ID
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - messageId
        - fileId
        - createdAt
...
paths:
...
  /messages:
    get:
      tags:
        - Messages
      summary: List messages
      description: Retrieve messages for a project
      parameters:
        - name: id
          in: query
          description: Message ID for single message retrieval
          schema:
            type: integer
        - name: projectId
          in: query
          description: Project ID (required for listing messages)
          schema:
            type: integer
        - name: pinned
          in: query
          description: Filter by pinned status
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of results (max 100)
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          description: Number of items to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Message'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Messages
      summary: Create message
      description: Create a new chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - role
                - content
              properties:
                projectId:
                  type: integer
                  description: Project ID
                role:
                  type: string
                  enum: [user, assistant]
                  description: Message sender role
                content:
                  type: string
                  description: Message content
                quotedMessageId:
                  type: integer
                  nullable: true
                  description: Quoted message ID
      responses:
        '201':
          description: Message created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Messages
      summary: Update message
      description: Update an existing message
      parameters:
        - name: id
          in: query
          required: true
          description: Message ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pinned:
                  type: boolean
                  description: Pinned status
                content:
                  type: string
                  description: Message content
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Messages
      summary: Delete message
      description: Delete an existing message
      parameters:
        - name: id
          in: query
          required: true
          description: Message ID
          schema:
            type: integer
      responses:
        '200':
          description: Message deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deleted:
                    $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /message-context-files:
    get:
      tags:
        - Message Context Files
      summary: List message context files
      description: Retrieve file associations for a message
      parameters:
        - name: id
          in: query
          description: Context file association ID
          schema:
            type: integer
        - name: messageId
          in: query
          description: Message ID to retrieve context files
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MessageContextFile'
                  - type: array
                    items:
                      $ref: '#/components/schemas/MessageContextFile'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Context file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Message Context Files
      summary: Add message context file
      description: Associate a file with a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messageId
                - fileId
              properties:
                messageId:
                  type: integer
                  description: Message ID
                fileId:
                  type: integer
                  description: File ID
      responses:
        '201':
          description: Context file added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageContextFile'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Message Context Files
      summary: Remove message context file
      description: Remove file association from a message
      parameters:
        - name: id
          in: query
          required: true
          description: Context file association ID
          schema:
            type: integer
      responses:
        '200':
          description: Context file removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deletedRecord:
                    $ref: '#/components/schemas/MessageContextFile'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Context file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...